+++ src/streetworks/common/__init__.py
@@ -56,6 +56,7 @@ from .from_mallorca import from_mallorca
 from .from_marousi import from_marousi_street
 from .from_milano import from_milano
 from .from_monaghan import from_monaghan_road
+from .from_na511 import from_na511
 from .from_nrn import from_nrn
 from .from_nsw_livetraffic import from_nsw_livetraffic
 from .from_nvdb import from_nvdb
@@ -208,4 +209,5 @@ __all__ = [
     "from_departement_roadworks",
     "from_dc",
     "from_quebec",
+    "from_na511",
 ]